requirements.txt
asttokens==2.4.1
attrs==23.2.0
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
datatest==0.11.1
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
h11==0.14.0
idna==3.6
ipython==8.18.1
jedi==0.19.1
matplotlib-inline==0.1.6
numpy==1.26.3
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.4
parso==0.8.3
pexpect==4.9.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
selenium==4.16.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
traitlets==5.14.0
trio==0.23.2
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==1.26.6
wcwidth==0.2.12
webdriver-manager==4.0.1
wsproto==1.2.0
